+++ src/main-curriculum/components/custom-footer.tsx
@@ -1,7 +1,11 @@
-export const CustomFooter = () => {
+interface Props {
+  name: string;
+}
+
+export const CustomFooter = ({ name }: Props) => {
   return (
-    <footer className="p-4 text-center bg-red-100 border-t-2 border-t-gray-50">
-      Custom header footer
+    <footer className="p-4 text-center text-white bg-blue-300 border-t-2 border-t-gray-50">
+      &copy; {name} (2025)
     </footer>
   );
 };