+++ src/main-curriculum/main-curriculum.tsx
@@ -1,14 +1,29 @@
+import { useEffect, useState } from "react";
+
 import {
   CustomAside,
   CustomFooter,
   CustomHeader,
   CustomMain,
 } from "./components";
+import type { PersonalInfoResponse } from "./interfaces";
+import { getPersonalInfo } from "./endpoints";
 
 export const MainCurriculum = () => {
+  const [personalInfo, setPersonalInfo] = useState<PersonalInfoResponse | null>(
+    null
+  );
+
+  useEffect(() => {
+    getPersonalInfo().then((resp: PersonalInfoResponse) => {
+      setPersonalInfo(resp);
+    });
+  }, []);
+
   return (
     <>
       <CustomHeader />
+
       <div className="flex flex-col md:flex-row">
         <div className="bg-orange-50 md:basis-1/4 md:border-r-2 md:border-r-gray-100">
           <CustomAside />
@@ -17,7 +32,8 @@ export const MainCurriculum = () => {
           <CustomMain />
         </div>
       </div>
-      <CustomFooter />
+
+      {personalInfo && <CustomFooter name={personalInfo.name} />}
     </>
   );
 };